Standardize CLI options and improve opening book configuration
- Rename CLI options for consistency (--leaves-path, --roots-*)
- Change opening book thresholds from double to int with better values
- Add UCT/PUCT bias factor providers to module configuration

+++ src/main/java/com/joansala/game/chess/Chess.java
@@ -161,11 +161,11 @@ public final class Chess {
     // Openings book
     // -------------------------------------------------------------------
 
-    /** Minimum score for an opening move to be chosen */
-    public static final double ROOT_THRESHOLD = -250.0D;
+    /** Never play book moves with a score lower than this */
+    public static final int ROOT_THRESHOLD = -250;
 
-    /** Threshold on the highest opening move reward */
-    public static final double ROOT_DISTURBANCE = 100.0D;
+    /** Play book moves within this range of the best move score */
+    public static final int ROOT_DISTURBANCE = 100;
 
     // -------------------------------------------------------------------
     // Board definitions
